+++ javascript/js-structure_shapes/components/Square/Square.js
@@ -1,0 +1,9 @@
+import { getRandomColor } from "../../utils/randomColor.js";
+export default function Square() {
+  const square = document.createElement("div");
+  square.classList.add("square");
+  square.addEventListener("click", () => {
+    square.style.backgroundColor = getRandomColor();
+  });
+  return square;
+}
