+++ javascript/js-structure_shapes/index.html
@@ -8,7 +8,7 @@
     />
     <title>HTML & CSS & JS</title>
     <link rel="stylesheet" href="./styles/styles.css" />
-    <script src="./index.js" defer></script>
+    <script type="module" src="./index.js" defer></script>
   </head>
   <body id="root"></body>
 </html>
